# Conflicts:
#	forgerock-openbanking-uk-aspsp-common/src/main/java/com/forgerock/openbanking/common/services/openbanking/event/EventValidationService.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-gateway/forgerock-openbanking-uk-aspsp-rs-gateway-server/src/main/java/com/forgerock/openbanking/aspsp/rs/api/payment/v3_1_3/domesticpayments/DomesticPaymentsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-gateway/forgerock-openbanking-uk-aspsp-rs-gateway-server/src/main/java/com/forgerock/openbanking/aspsp/rs/api/payment/v3_1_3/domesticscheduledpayments/DomesticScheduledPaymentsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-gateway/forgerock-openbanking-uk-aspsp-rs-gateway-server/src/main/java/com/forgerock/openbanking/aspsp/rs/api/payment/v3_1_3/domesticstandingorders/DomesticStandingOrdersApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-gateway/forgerock-openbanking-uk-aspsp-rs-gateway-server/src/main/java/com/forgerock/openbanking/aspsp/rs/api/payment/v3_1_3/file/FilePaymentsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-gateway/forgerock-openbanking-uk-aspsp-rs-gateway-server/src/main/java/com/forgerock/openbanking/aspsp/rs/api/payment/v3_1_3/internationalpayments/InternationalPaymentsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-gateway/forgerock-openbanking-uk-aspsp-rs-gateway-server/src/main/java/com/forgerock/openbanking/aspsp/rs/api/payment/v3_1_3/internationalscheduledpayments/InternationalScheduledPaymentsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-gateway/forgerock-openbanking-uk-aspsp-rs-gateway-server/src/main/java/com/forgerock/openbanking/aspsp/rs/api/payment/v3_1_3/internationalstandingorders/InternationalStandingOrdersApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-gateway/forgerock-openbanking-uk-aspsp-rs-gateway-server/src/main/java/com/forgerock/openbanking/aspsp/rs/api/payment/v3_1_5/domesticpayments/DomesticPaymentConsentsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-gateway/forgerock-openbanking-uk-aspsp-rs-gateway-server/src/main/java/com/forgerock/openbanking/aspsp/rs/api/payment/v3_1_5/domesticpayments/DomesticPaymentsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-gateway/forgerock-openbanking-uk-aspsp-rs-gateway-server/src/main/java/com/forgerock/openbanking/aspsp/rs/api/payment/v3_1_5/domesticscheduledpayments/DomesticScheduledPaymentsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-gateway/forgerock-openbanking-uk-aspsp-rs-gateway-server/src/main/java/com/forgerock/openbanking/aspsp/rs/api/payment/v3_1_5/domesticstandingorders/DomesticStandingOrdersApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-gateway/forgerock-openbanking-uk-aspsp-rs-gateway-server/src/main/java/com/forgerock/openbanking/aspsp/rs/api/payment/v3_1_5/file/FilePaymentsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-gateway/forgerock-openbanking-uk-aspsp-rs-gateway-server/src/main/java/com/forgerock/openbanking/aspsp/rs/api/payment/v3_1_5/internationalpayments/InternationalPaymentsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-gateway/forgerock-openbanking-uk-aspsp-rs-gateway-server/src/main/java/com/forgerock/openbanking/aspsp/rs/api/payment/v3_1_5/internationalscheduledpayments/InternationalScheduledPaymentsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-gateway/forgerock-openbanking-uk-aspsp-rs-gateway-server/src/main/java/com/forgerock/openbanking/aspsp/rs/api/payment/v3_1_5/internationalstandingorders/InternationalStandingOrdersApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/main/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/event/EventsHelper.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/main/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/event/v3_0/CallbackUrlsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/main/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/event/v3_1/CallbackUrlsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/main/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/event/v3_1_1/CallbackUrlsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/main/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/event/v3_1_2/callbackurl/CallbackUrlsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/main/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/event/v3_1_2/eventsubscription/EventSubscriptionApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/main/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/event/v3_1_3/callbackurl/CallbackUrlsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/main/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/event/v3_1_4/callbackurl/CallbackUrlsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/main/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/event/v3_1_5/callbackurl/CallbackUrlsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/main/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/event/v3_1_6/callbackurl/CallbackUrlsApiController.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/test/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/event/v3_0/CallbackUrlsApiControllerIT.java
#	forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/test/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/event/v3_1_2/eventsubscription/EventSubscriptionApiControllerIT.java

+++ forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/main/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/event/v3_0/CallbackUrlsApiController.java
@@ -20,16 +20,16 @@
  */
 package com.forgerock.openbanking.aspsp.rs.store.api.openbanking.event.v3_0;
 
-import com.forgerock.openbanking.aspsp.rs.store.api.openbanking.event.EventsHelper;
 import com.forgerock.openbanking.aspsp.rs.store.repository.TppRepository;
 import com.forgerock.openbanking.aspsp.rs.store.repository.events.CallbackUrlsRepository;
-import com.forgerock.openbanking.common.model.openbanking.domain.event.FRCallbackUrlData;
+import com.forgerock.openbanking.common.conf.discovery.ResourceLinkService;
 import com.forgerock.openbanking.common.model.openbanking.persistence.event.FRCallbackUrl;
 import com.forgerock.openbanking.common.model.version.OBVersion;
-import com.forgerock.openbanking.exceptions.OBErrorResponseException;
+import com.forgerock.openbanking.common.services.openbanking.event.EventResponseUtil;
 import com.forgerock.openbanking.model.Tpp;
 import io.swagger.annotations.ApiParam;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -37,10 +37,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import uk.org.openbanking.datamodel.event.OBCallbackUrl1;
-import uk.org.openbanking.datamodel.event.OBCallbackUrlResponse1;
-import uk.org.openbanking.datamodel.event.OBCallbackUrlResponseData1;
-import uk.org.openbanking.datamodel.event.OBCallbackUrlsResponse1;
-import uk.org.openbanking.datamodel.event.OBCallbackUrlsResponseData1;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
@@ -48,7 +44,6 @@ import java.security.Principal;
 import java.util.Collection;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 import static com.forgerock.openbanking.common.services.openbanking.converter.event.FRCallbackUrlConverter.toFRCallbackUrlData;
 
@@ -56,12 +51,24 @@ import static com.forgerock.openbanking.common.services.openbanking.converter.ev
 @Slf4j
 public class CallbackUrlsApiController implements CallbackUrlsApi {
 
-    private CallbackUrlsRepository callbackUrlsRepository;
-    private TppRepository tppRepository;
+    private final CallbackUrlsRepository callbackUrlsRepository;
+    private final TppRepository tppRepository;
+    private final EventResponseUtil eventResponseUtil;
+    private final ResourceLinkService resourceLinkService;
 
-    public CallbackUrlsApiController(CallbackUrlsRepository callbackUrlsRepository, TppRepository tppRepository) {
+    @Autowired
+    public CallbackUrlsApiController(CallbackUrlsRepository callbackUrlsRepository, TppRepository tppRepository, ResourceLinkService resourceLinkService) {
         this.callbackUrlsRepository = callbackUrlsRepository;
         this.tppRepository = tppRepository;
+        this.resourceLinkService = resourceLinkService;
+        this.eventResponseUtil = new EventResponseUtil(this.resourceLinkService, OBVersion.v3_0, false);
+    }
+
+    public CallbackUrlsApiController(CallbackUrlsRepository callbackUrlsRepository, TppRepository tppRepository,ResourceLinkService resourceLinkService, EventResponseUtil eventResponseUtil) {
+        this.callbackUrlsRepository = callbackUrlsRepository;
+        this.tppRepository = tppRepository;
+        this.resourceLinkService = resourceLinkService;
+        this.eventResponseUtil = eventResponseUtil;
     }
 
     @Override
@@ -88,18 +95,18 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
             HttpServletRequest request,
 
             Principal principal
-    ) throws OBErrorResponseException {
+    ) {
         log.debug("Create new callback URL: {} for client: {}", obCallbackUrl1Param, clientId);
 
         // Check if callback URL already exists for TPP
         // https://openbanking.atlassian.net/wiki/spaces/DZ/pages/645367055/Event+Notification+API+Specification+-+v3.0#EventNotificationAPISpecification-v3.0-POST/callback-urls
+        // A TPP must only create a callback-url on one version
         final Optional<Tpp> isTpp = Optional.ofNullable(tppRepository.findByClientId(clientId));
 
         if (!isTpp.isPresent()) {
             log.warn("No TPP found for client id '{}'", clientId);
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("TPP not found");
         }
-
 
         final Collection<FRCallbackUrl> byClientId = callbackUrlsRepository.findByTppId(isTpp.get().getId());
         final boolean urlExists = byClientId.stream()
@@ -118,7 +125,7 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
 
         return ResponseEntity
                 .status(HttpStatus.CREATED)
-                .body(packageResponse(frCallbackUrl));
+                .body(eventResponseUtil.packageResponse(frCallbackUrl));
     }
 
     @Override
@@ -142,7 +149,14 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
         return Optional.ofNullable(tppRepository.findByClientId(clientId))
                 .map(Tpp::getId)
                 .map(id -> callbackUrlsRepository.findByTppId(id))
-                .map(urls -> ResponseEntity.ok(packageResponse(urls)))
+                // A TPP must only create a callback-url on one version and then there must be only one and we take only the first
+                .map(urls -> {
+                    if (urls.isEmpty()) {
+                        log.warn("No CallbackURL found for client id '{}'", clientId);
+                        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+                    }
+                    return ResponseEntity.ok(eventResponseUtil.packageResponse(urls));
+                })
                 .orElseGet(() ->
                         {
                             log.warn("No TPP found for client id '{}'", clientId);
@@ -178,14 +192,18 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
             HttpServletRequest request,
 
             Principal principal
-    ) throws OBErrorResponseException {
+    ) {
         final Optional<FRCallbackUrl> byId = callbackUrlsRepository.findById(callbackUrlId);
 
         if (byId.isPresent()) {
             FRCallbackUrl frCallbackUrl = byId.get();
-            frCallbackUrl.setCallbackUrl(toFRCallbackUrlData(obCallbackUrl1Param));
-            callbackUrlsRepository.save(frCallbackUrl);
-            return ResponseEntity.ok(packageResponse(frCallbackUrl));
+            if(eventResponseUtil.IsAllowedAccessResourceFromApiVersionInstanced(frCallbackUrl.getCallbackUrl().getVersion())){
+                frCallbackUrl.setCallbackUrl(toFRCallbackUrlData(obCallbackUrl1Param));
+                callbackUrlsRepository.save(frCallbackUrl);
+                return ResponseEntity.ok(eventResponseUtil.packageResponse(frCallbackUrl));
+            }else{
+                return ResponseEntity.status(HttpStatus.CONFLICT).body("Callback URL: '" + callbackUrlId + "' can't be update via an older API version.");
+            }
         } else {
             // Spec isn't clear on if we should
             // 1. Reject a PUT for a resource id that does not exist
@@ -215,40 +233,18 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
             HttpServletRequest request,
 
             Principal principal
-    ) throws OBErrorResponseException {
+    ) {
         final Optional<FRCallbackUrl> byId = callbackUrlsRepository.findById(callbackUrlId);
         if (byId.isPresent()) {
-            log.debug("Deleting callback url: {}", byId.get());
-            callbackUrlsRepository.deleteById(callbackUrlId);
-            return ResponseEntity.noContent().build();
+            if(eventResponseUtil.IsAllowedAccessResourceFromApiVersionInstanced(byId.get().getCallbackUrl().getVersion())){
+                log.debug("Deleting callback url: {}", byId.get());
+                callbackUrlsRepository.deleteById(callbackUrlId);
+                return ResponseEntity.noContent().build();
+            }else {
+                return ResponseEntity.status(HttpStatus.CONFLICT).body("Callback URL: '" + callbackUrlId + "' can't be delete via an older API version.");
+            }
         } else {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Callback URL: '" + callbackUrlId + "' can't be found");
         }
-    }
-
-    protected OBCallbackUrlsResponse1 packageResponse(Collection<FRCallbackUrl> frCallbackUrls) {
-        return new OBCallbackUrlsResponse1()
-                .data(new OBCallbackUrlsResponseData1()
-                        .callbackUrl(
-                                frCallbackUrls.stream()
-                                        .filter(EventsHelper.matchingVersion(OBVersion.v3_0))
-                                        .map(this::toOBCallbackUrlResponseData1)
-                                        .collect(Collectors.toList())
-                        )
-                );
-    }
-
-    private OBCallbackUrlResponse1 packageResponse(FRCallbackUrl frCallbackUrl) {
-        return new OBCallbackUrlResponse1()
-                .data(toOBCallbackUrlResponseData1(frCallbackUrl));
-    }
-
-
-    protected OBCallbackUrlResponseData1 toOBCallbackUrlResponseData1(FRCallbackUrl frCallbackUrl) {
-        final FRCallbackUrlData data = frCallbackUrl.getCallbackUrl();
-        return new OBCallbackUrlResponseData1()
-                .callbackUrlId(frCallbackUrl.getId())
-                .url(data.getUrl())
-                .version(data.getVersion());
     }
 }
